Renamed NotFound.js to 404.js
